+++ interop-testing/src/main/java/io/grpc/testing/integration/TestServiceServer.java
@@ -146,9 +146,7 @@ public class TestServiceServer {
         .maxMessageSize(AbstractInteropTest.MAX_MESSAGE_SIZE)
         .addService(ServerInterceptors.intercept(
             new TestServiceImpl(executor),
-            TestUtils.echoRequestHeadersInterceptor(Util.METADATA_KEY),
-            TestUtils.echoRequestMetadataInHeaders(Util.ECHO_INITIAL_METADATA_KEY),
-            TestUtils.echoRequestMetadataInTrailers(Util.ECHO_TRAILING_METADATA_KEY)))
+            TestServiceImpl.interceptors()))
         .build().start();
   }
